+++ Editor/MCPServerMethods.Component.cs
@@ -89,18 +89,31 @@ namespace UnityMCP.Editor
                 case SerializedPropertyType.Boolean: prop.boolValue = (bool)value; break;
                 case SerializedPropertyType.Float: prop.floatValue = (float)value; break;
                 case SerializedPropertyType.String: prop.stringValue = value.ToString(); break;
+                default: ApplyComplexValueToProperty(prop, value); break;
+            }
+        }
+
+        private static void ApplyComplexValueToProperty(SerializedProperty prop, JToken value)
+        {
+            switch (prop.propertyType)
+            {
                 case SerializedPropertyType.Vector3: prop.vector3Value = ParseVector3(value, prop.vector3Value); break;
-                case SerializedPropertyType.Color: 
+                case SerializedPropertyType.Color:
                     if (ColorUtility.TryParseHtmlString(value.ToString(), out Color color)) prop.colorValue = color;
                     break;
-                case SerializedPropertyType.Enum: 
-                    if (value.Type == JTokenType.Integer) prop.enumValueIndex = (int)value;
-                    else
-                    {
-                        int index = Array.IndexOf(prop.enumDisplayNames, value.ToString());
-                        if (index >= 0) prop.enumValueIndex = index;
-                    }
+                case SerializedPropertyType.Enum:
+                    ApplyEnumValue(prop, value);
                     break;
+            }
+        }
+
+        private static void ApplyEnumValue(SerializedProperty prop, JToken value)
+        {
+            if (value.Type == JTokenType.Integer) prop.enumValueIndex = (int)value;
+            else
+            {
+                int index = Array.IndexOf(prop.enumDisplayNames, value.ToString());
+                if (index >= 0) prop.enumValueIndex = index;
             }
         }